+++ src/config.cpp
@@ -587,6 +587,7 @@ namespace pinpoint {
         {"Http.Server.RecordRequestCookie", REF(http.server.rec_request_cookie), RELOAD, env::HTTP_SERVER_RECORD_REQUEST_COOKIE},
         {"Http.Server.RecordResponseHeader", REF(http.server.rec_response_header), RELOAD, env::HTTP_SERVER_RECORD_RESPONSE_HEADER},
         {"Http.Server.ProxyUserHeaderNames", REF(http.server.proxy_user_header_names), RELOAD, env::HTTP_SERVER_PROXY_USER_HEADER_NAMES},
+        {"Http.Server.ProxyHeaderEnable", REF(http.server.proxy_header_enable), RELOAD, env::HTTP_SERVER_PROXY_HEADER_ENABLE},
         {"Http.Client.RecordRequestHeader", REF(http.client.rec_request_header), RELOAD, env::HTTP_CLIENT_RECORD_REQUEST_HEADER},
         {"Http.Client.RecordRequestCookie", REF(http.client.rec_request_cookie), RELOAD, env::HTTP_CLIENT_RECORD_REQUEST_COOKIE},
         {"Http.Client.RecordResponseHeader", REF(http.client.rec_response_header), RELOAD, env::HTTP_CLIENT_RECORD_RESPONSE_HEADER},